Fix Airtable Cells for T1, W1P1, W1P2
The airtable cells need to be atomic in the sense that they can be run at any point. The try/catch blocks enable the cell to be rendered even if some variables are missing. Expected behavior: should not throw a NameError, Airtable should render properly, default values are filled in.